Train in fp32 at lr 1e-4; count non-finite losses instead of dropping them
With the HRNet weights actually loaded, fp16 training NaNs: on a
16-drawing overfit loss_mask and loss_dice went NaN for every drawing by
epoch 5 and never recovered; computing mask logits and mask losses in
float32 (kept -- it is correct regardless) only delayed it to epoch 14.
The same run in fp32 stayed finite with loss 237 -> 120 over 20 epochs.
Both 3060 configs switch to fp16: False until the overflowing op is found.

train.py skipped NaN losses silently -- 'loss > 0' is False for NaN -- so
a NaN run looked like a run that was not learning. They are now counted,
logged, and abort the run past max_skip_frac.

lr 2e-4 -> 1e-4: Stage-1 loss climbed through epoch 2 at 2e-4; a
single-drawing overfit converges cleanly at 1e-4.

Stage 1 is cut to 6 epochs as a warm start for the joint run, which
already carries FloorPlanCAD at weight 0.35. The first Stage-1 run (image
branch never loaded) is discarded; its log is kept in
work_dirs/_stage1_broken_hrnet/.

+++ svgnet/model/criterion.py
@@ -225,7 +225,7 @@ class SetCriterion(nn.Module):
         tgt_idx = self._get_tgt_permutation_idx(indices)
         src_masks = outputs["pred_masks"]
         
-        src_masks = src_masks[src_idx]
+        src_masks = src_masks[src_idx].float()
         masks = [t["masks"].transpose(0,1).unsqueeze(0) for t in targets]
         target_masks = torch.cat(masks,dim=0)
